+++ backend/src/pragma/__init__.py
@@ -10,4 +10,4 @@ Returns:
 Raises:
     None.
 """
-__version__ = "0.13.1"
+__version__ = "0.14.0"